reap after SIGKILL again
Strictly speaking not needed, but I think it may case the filesystems
to remain EBUSY.

Also, when running in a container we don't want to leak zombies to the
outside (found out by skarnet).

We need the alarm and blocking waitpid, since it's possible (in theory)
no process has exited between the kill and first call of waitpid(-1, 0,
WNOHANG).