Linux: work around processes stuck in state D blocking shutdown
This blog post explains how to create TASK_UNINTERRUPTIBLE processes
that cannot be killed:
https://chrisdown.name/2024/02/05/reliably-creating-d-state-processes-on-demand.html

This usually only happens due to kernel or hardware bugs, but it still
should not stop nitro from rebooting.

Therefore, we need to run "kill -9 -1" in an exec'ed process so pid 1
is not blocked due to kernel behavior.

(Potentially needed on other OS too, untested there.)