delete stale control socket on start
We do not unconditionally want to remove the control socket since
someone could accidentally start nitro while it's running as init
already.  Only unlink the socket when it has no listener.

We connect with the wrong socket type so we don't cause a connection
in case nitro is running already.